requirements.txt
streamlit
openai
pytidb

python-dotenv